svg-template-api/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==36.0.2
Flask==3.1.0
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyMySQL==1.1.1
SQLAlchemy==2.0.40
typing_extensions==4.14.0
Werkzeug==3.1.3
cairosvg==2.7.1
Pillow==10.4.0
lxml==5.3.0

